Progess on refactoring population, genotype and manager interactions
Split up population and move some of its responsibilities to a manager class.
Prepared some of the stubs for unit testing later on the evolution package.
Refactored the recovery procedure done normally in the manager, and moved it to a separate set of functions.
Added a robot identifier which can be globally called in order to update the genotype id's.