Refactoring most of the evolution package
Refactored almost all classes except for the implementation of the speciation package.
First of the generation and robot identifiers are based on a sequential singleton, granting it better distribution of accessing the robot and generational index.
Seperate loading from the class object.
Renamed PopulationSpeciated to Speciation in order to avoid the longer name without adding to the population class since the amount of generalization is still to be decided after reworking the speciation package.

Moved specific implementation of behavioral measures to its class instead of the recovering function.